Move keep_larger to iterables utils
